// src/app/services/interfaces/api.interfaces.ts
export interface FecthAPIKey {
    token: string;
}

export interface OFund {
    tipo: string;
    nfund: NFund[];
  }
  
export interface NFund {
    nombre: string;
    fund: Fundamento[];
  }

/* export interface FetchAllTramites {
    homoclave:        string;
    nombre:           string;
    modalidad:        string;
    sujetoObligado:   string;
    fechaPublicacion: string;
} */


export interface FetchAllTramites {
    homoclave:        string;
    nombre:           string;
    modalidad:        string;
    sujetoObligado:   string;
    descripcion:      string;
    fechaPublicacion: string;
}