// src/app/services/interfaces/dep.api.interface.ts
// Generated by https://quicktype.io

export interface DependenciasAPITs {
    data:  Dependencia[];
    totalItems: number;
    pageNumber: number;
    rows:       number;
}

export interface Dependencia {
    id:           number;
    codigo:       null;
    nombre:       string;
    siglas:       string;
    naturaleza:   Naturaleza;
    idPadre:      number;
    esAMR:        null;
    amrId:        null;
    amrAlternaId: null;
    sector_CatId: null;
    activo:       boolean;
}

/* https://articulo.mercadolibre.com.mx/MLM-856632012-laptop-lenovo-ideapad-500-14isk-w-16gb-ssd-120gb-_JM#position=1&search_layout=stack&type=item&tracking_id=19ac9be6-3f0f-47be-8b01-fff7ee68fc37
 */
export enum Naturaleza {
    Empty = "-",
    Null = "null",
}