// src/app/services/interfaces/tram.api.interface.ts
/* export interface FindTram {
    data:       DataTamite[];
    totalItems: number;
    pageNumber: number;
    rows:       number;
}

export interface DataTamite {
    id:            number;
    nombre:        string;
    homoclave:     string;
    categoria:     string;
    ordenamientos: Ordenamiento[];
}

export interface Ordenamiento {
    id:       number;
    nombre:   string;
    articulo: string;
    fraccion: null | string;
    insiso:   null | string;
    parrafo:  null | string;
    numero:   null | string;
    letra:    null | string;
    otro:     null | string;
}
 */

// Generated by https://quicktype.io

export interface FindTram {
    data:       DataTamite[];
    totalItems: number;
    pageNumber: number;
    rows:       number;
}

export interface DataTamite {
    id:                   number;
    nombre:               string;
    homoclave:            string;
    categoria:            string;
    modalidad:            string;
    sujetoObligado:       string;
    descripcionCiudadana: string;
    traFechaModificacion: string;
    ordenamientos:        Ordenamiento[];
}

export interface Ordenamiento {
    id:       number;
    nombre:   string;
    articulo: string;
    fraccion: null | string;
    insiso:   null | string;
    parrafo:  null;
    numero:   null;
    letra:    null;
    otro:     null;
}